+++ Cesium3DTilesSelection/src/TilesetContentManager.cpp
@@ -2035,15 +2035,9 @@ void TilesetContentManager::propagateTilesetContentLoaderResult(
   this->_requestHeaders = std::move(result.requestHeaders);
   this->_pLoader = std::move(result.pLoader);
   this->_pRootTile = std::move(result.pRootTile);
-  this->_voxelExtension = std::move(result.voxelExtension);
 
   this->_overlayCollection.setLoadedTileEnumerator(
       LoadedTileEnumerator(this->_pRootTile.get()));
   this->_pLoader->setOwner(*this);
-}
-
-const Cesium3DTiles::ExtensionContent3dTilesContentVoxels*
-TilesetContentManager::getVoxelExtension() const noexcept {
-  return this->_voxelExtension ? &(*this->_voxelExtension) : nullptr;
 }
 } // namespace Cesium3DTilesSelection